requirements.txt
streamlit
streamlit-chat
PyPDF2
langchain
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
python-dotenv
nemoguardrails
nest-asyncio
openai
sentence-transformers
langchain-groq
